requirements.txt
django==3.1.3
djangorestframework==3.12.2
django-redis==4.12.1
redis==3.5.3
celery==5.0.2
django-celery-beat==2.2.0

psycopg2==2.8.6
pyyaml==5.3.1

whitenoise==5.2.0
gunicorn==20.0.4

black==20.8b1
flake8==3.8.4
